Add files via upload
Lufthansa Cargo Flight Schedule visualised.ipynb

https://blog.hugo-larcher.com/flight-data-visualisation-with-pandas-and-matplotlib-ebbd13038647



We want to replicate the code provided in the blog post for visualizing flight data using Pandas and Matplotlib, but with the Lufthansa Cargo (LCAG) Flight Schedule available at https://www.lufthansa-cargo.com/de/network/schedule-routings. To achieve this, we need to preprocess the LCAG data to match the format of the Air France data. This involves obtaining the total number of flights per week for a specific routing and collecting the longitude and latitude information for mapping the arrival and departure airports on a map.

To accomplish this, we plan to join the LCAG flight schedule with a dataset (https://datahub.io/core/airport-codes) containing worldwide airport codes along with their corresponding coordinates (longitude and latitude). After the join operation, we can filter out unnecessary columns and proceed to map the schedule on a geographical map for visualization.